requirements.txt
Pillow==8.2.0
numpy==1.20.3
pycoral==0.0.7